+++ Packages/PrefsGUI.RapidGUI/Runtime/PrefsMinMaxExtension.cs
@@ -4,20 +4,17 @@ namespace PrefsGUI.RapidGUI
 {
     public static class PrefsMinMaxExtension
     {
-        public static bool DoGUISlider<T, TMinMax>(this PrefsMinMax<T, TMinMax> prefs, string label = null)
-            where TMinMax : MinMax<T>, new()
+        public static bool DoGUISlider<T>(this PrefsMinMax<T> prefs, string label = null)
         {
             return prefs.DoGUISlider(prefs.defaultMin, prefs.defaultMax, label);
         }
 
-        public static bool DoGUISlider<T, TMinMax>(this PrefsMinMax<T, TMinMax> prefs, T max, string label = null)
-            where TMinMax : MinMax<T>, new()
+        public static bool DoGUISlider<T>(this PrefsMinMax<T> prefs, T max, string label = null)
         {
             return prefs.DoGUISlider(prefs.defaultMin, max, label);
         }
         
-        public static bool DoGUISlider<T, TMinMax>(this PrefsMinMax<T, TMinMax> prefs, T rangeMin, T rangeMax, string label = null)
-            where TMinMax : MinMax<T>, new()
+        public static bool DoGUISlider<T>(this PrefsMinMax<T> prefs, T rangeMin, T rangeMax, string label = null)
         {
             return prefs.DoGUIStandard((v) =>
             {